feat:배송중 함수 ì구현

+++ back/src/main/java/org/back/back/domain/order/service/OrderService.java
@@ -5,6 +5,8 @@ import org.back.back.domain.order.entity.Order;
 import org.back.back.domain.order.repository.OrderRepository;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -18,12 +20,27 @@ public class OrderService {
 
     public Order modify(int id, int quantity){
         Order order = orderRepository.findById(id).orElseThrow();
+        if(isShipping(order)){
+            throw new RuntimeException("배송중인 주문은 수정할 수 없습니다");
+        }
         order.modify(quantity);
         return order;
     }
 
     public void delete(int id){
         Order order = orderRepository.findById(id).orElseThrow();
+        if(isShipping(order)){
+            throw new RuntimeException("배송중인 주문은 삭제할 수 없습니다");
+        }
         orderRepository.delete(order);
+    }
+
+    // 배송중을 확인하는 함수
+    // 예) 8/24 14:00 ~ 8/25 13:59:59 주문 -> deliveryDate = 8/25 -> 8/25 14:00부터 배송중
+    // 배송중 -> 수정,삭제 불가
+    private boolean isShipping(Order order){
+        LocalDateTime now = LocalDateTime.now();
+        LocalDate deliberyDate = order.getDeliveryDate();
+        return !now.isBefore(deliberyDate.atTime(14,0));
     }
 }
